requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
docutils==0.20.1
fonttools==4.50.0
idna==3.6
importlib_metadata==7.1.0
jaraco.classes==3.4.0
jaraco.context==4.3.0
jaraco.functools==4.0.0
keyring==25.1.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplot==0.1.9
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.2.0
nh3==0.2.17
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pkginfo==1.10.0
Pygments==2.17.2
pyloco==0.0.139
pyparsing==3.1.2
PyQt6==6.6.1
PyQt6-Qt6==6.6.2
PyQt6-sip==13.6.0
pyqtgraph==0.13.4
pyserial==3.5
python-dateutil==2.9.0.post0
readme_renderer==43.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
SimpleWebSocketServer==0.1.2
six==1.16.0
twine==5.0.0
typing==3.7.4.3
urllib3==2.2.1
ushlex==0.99.1
websocket-client==1.7.0
zipp==3.18.1
